feat: reporte de ventas por técnico | nueva card en vista de reportes | manejo de saldos sin ISV y saldos negativos en bitácoras | validación de políticas por rol | endpoint para actualizar calificación

+++ app/api/usuarios/[id]/route.ts
@@ -1,9 +1,10 @@
 import { z } from "zod";
 import { NextResponse } from "next/server";
-import { UsuarioService } from "@/app/services/usuarioService";
-import { ResponseDto } from "@/app/common/dtos/response.dto";
 import { CrearUsuarioDto } from "@/app/dtos/usuario.dto";
+import { ResponseDto } from "@/app/common/dtos/response.dto";
+import { UsuarioService } from "@/app/services/usuarioService";
 import { GeneralUtils } from "@/app/common/utils/general.utils";
+import { obtenerPayloadSesion } from "@/app/common/utils/session.utils";
 
 const EditarUsuarioDto = CrearUsuarioDto.omit({ password: true }).partial();
 type EditarUsuarioDto = z.infer<typeof EditarUsuarioDto>;
@@ -56,6 +57,11 @@ export async function DELETE(req: Request, { params }: { params: Promise<{ id: s
     
     try {
         
+        const payload = await obtenerPayloadSesion();
+        if (payload.rol !== "admin") {
+            return NextResponse.json(new ResponseDto(403, "No tienes permiso para acceder a esta información"));
+        }
+
         const idParams = (await params).id;
         const id = GeneralUtils.validarIdParam(idParams);
         const usuarioEliminado = await UsuarioService.eliminarUsuario(id);
